ios: set audio session category, notify other apps when they can resume
The `AudioRecorder` class has already sets the audio session category
to `playAndRecord` but since playback of voice messages is handled by
the `AudioPlayer` class it did not set the audio session for voice
messages. In addition after recording it should probably notify the
audio session that it has finished recording and set the category back
to `ambient`.